+++ srcs/ft_input_cycle.c
@@ -6,7 +6,7 @@
 /*   By: mrantil <mrantil@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/20 11:46:24 by mrantil           #+#    #+#             */
-/*   Updated: 2022/11/11 12:23:34 by mrantil          ###   ########.fr       */
+/*   Updated: 2022/11/11 15:46:41 by mrantil          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,14 +37,14 @@ static void ft_restart_cycle(t_term *t)
 	t->total_row = 0;
 	t->c_row = t->total_row;
 	ft_add_nl_last_row(t, 0);
-	write(1, PROMPT, t->prompt_len);
+	write(1, PROMPT, (size_t)t->prompt_len);
 	ft_setcursor(t->c_col, t->c_row + t->start_row);
 }
 
 void	ft_input_cycle(t_term *t)
 {
 	ft_add_nl_last_row(t, 0);
-	write(1, PROMPT, t->prompt_len);
+	write(1, PROMPT, (size_t)t->prompt_len);
 	while (t->ch != -1)
 	{
 		t->ch = ft_get_input();
